+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 - Optimized code and cleaned clutter.
 - Reduced file size and reworked update mechanism for terrors.
 - Renamed some internal concepts.
+- tontrack.me's label in Options is now a hyperlink for ease of access.
 
 ## [alpha-0.7.7] - 2025.04.10